+++ packages/viz/src/modules/print/BasicsPrintView.js
@@ -309,7 +309,7 @@ class BasicsPrintView extends PrintView {
     );
 
     const { averageGlucose } = _.get(this.data.stats, "averageGlucose.data.raw", {});
-    this.renderSimpleStat(t("Average BG"), formatDecimalNumber(averageGlucose), "\nmg/dL");
+    this.renderSimpleStat(t("Average BG"), formatDecimalNumber(averageGlucose), `\n${this.bgUnits}`);
 
     const { glucoseManagementIndicator } = _.get(this.data.stats, "glucoseManagementIndicator.data.raw", {});
     this.renderSimpleStat(t("GMI ({{bgSourceLabel}})"), formatDecimalNumber(glucoseManagementIndicator, 1), "%");